+++ project/bodquest_2023/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'application/usecase/firebase_options.dart';
+import 'infrastructure/datasource/firebase_options.dart';
 
 import 'presentation/page/myapp_page.dart';
